requirements.txt
flask
pypdf2
google-genai
python-dotenv
gunicorn